Time: 11 ms (86.05%), Space: 50.2 MB (45.56%) - LeetHub

+++ 0387-first-unique-character-in-a-string/0387-first-unique-character-in-a-string.java
@@ -8,7 +8,7 @@ class Solution {
             arr[ch-'a']+=1;
         
         for(int i=0; i<s.length(); i++)
-            if(arr[(int)s.charAt(i)-'a'] == 1) return i;
+            if(arr[s.charAt(i)-'a'] == 1) return i;
         
         return -1;
     }
